Break run and result-related logic into a separate component, update flow run results compnent
