+++ src/components/ArtifactCard.vue
@@ -8,10 +8,6 @@
         <div v-if="hasKey" class="artifact-card__title">
           {{ artifact.key }}
         </div>
-
-        <template v-if="hasRun && crumbs.length">
-          <p-bread-crumbs :crumbs="crumbs" class="artifact-card__bread-crumbs" :class="classes.breadCrumbs" />
-        </template>
       </header>
 
       <div class="artifact-card__summary-container" :class="classes.summaryContainer">
@@ -33,10 +29,7 @@
 </template>
 
 <script lang="ts" setup>
-  import { BreadCrumbs } from '@prefecthq/prefect-design'
   import { computed, useSlots } from 'vue'
-  import { useRoute } from 'vue-router'
-  import { useFlowRun, useTaskRun, useWorkspaceRoutes } from '@/compositions'
   import { localization } from '@/localization'
   import { Artifact } from '@/models'
   import { formatDateTime } from '@/utilities'
@@ -47,40 +40,7 @@
   }>()
 
   const slots = useSlots()
-  const routes = useWorkspaceRoutes()
-  const route = useRoute()
-
-  const flowRunId = computed(() => props.artifact.flowRunId)
-  const taskRunId = computed(() => props.artifact.taskRunId)
-
-  const flowRun = useFlowRun(flowRunId)
-  const taskRun = useTaskRun(taskRunId)
-
-  const hasRun = computed(() => !!props.artifact.flowRunId || !!props.artifact.taskRunId)
   const hasKey = computed(() => !!props.artifact.key)
-
-  const viewingFlowRun = computed(() => flowRunId.value && route.fullPath.includes(flowRunId.value))
-  const viewingTaskRun = computed(() => taskRunId.value && route.fullPath.includes(taskRunId.value))
-
-  const crumbs = computed<BreadCrumbs>(() => {
-    const internalCrumbs: BreadCrumbs = []
-
-    if (flowRun.value && !viewingFlowRun.value) {
-      internalCrumbs.push({
-        text: flowRun.value.name ?? '',
-        to: routes.flowRun(flowRun.value.id),
-      })
-    }
-
-    if (taskRun.value && !viewingTaskRun.value) {
-      internalCrumbs.push({
-        text: taskRun.value.name ?? '',
-        to: routes.taskRun(taskRun.value.id),
-      })
-    }
-
-    return internalCrumbs
-  })
 
   const classes = computed(() => {
     return {
@@ -95,9 +55,6 @@
       },
       subheader: {
         'artifact-card__subheader--condensed': props.condense,
-      },
-      breadCrumbs: {
-        'artifact-card__bread-crumbs--condensed': props.condense,
       },
       summaryContainer: {
         'artifact-card__summary-container--condensed': props.condense,
@@ -155,21 +112,6 @@
   text-xs
   text-foreground-200
   uppercase
-}
-
-.artifact-card__bread-crumbs { @apply
-  text-base
-  flex-nowrap
-  overflow-hidden
-}
-
-.artifact-card__bread-crumbs .p-bread-crumbs__crumb { @apply
-  whitespace-nowrap
-  overflow-ellipsis
-}
-
-.artifact-card__bread-crumbs--condensed { @apply
-  text-sm
 }
 
 .artifact-card__summary-container { @apply
