+++ src/components/index.ts
@@ -8,6 +8,7 @@ export { default as ArtifactDataResult } from './ArtifactDataResult.vue'
 export { default as ArtifactDataUnknown } from './ArtifactDataUnknown.vue'
 export { default as ArtifactDescription } from './ArtifactDescription.vue'
 export { default as ArtifactMenu } from './ArtifactMenu.vue'
+export { default as ArtifactCardResult } from './ArtifactCardResult.vue'
 export { default as ArtifactTimeline } from './ArtifactTimeline.vue'
 export { default as ArtifactTypeSelect } from './ArtifactTypeSelect.vue'
 export { default as BetaBadge } from './BetaBadge.vue'
